Move credential secret logic to CacheableSecretsManager
