+++ athena-federation-sdk/src/main/java/com/amazonaws/athena/connector/lambda/handlers/RecordHandler.java
@@ -115,6 +115,11 @@ public abstract class RecordHandler
         return secretsManager.resolveSecrets(rawString);
     }
 
+    protected String resolveWithDefaultCredentials(String rawString)
+    {
+        return secretsManager.resolveWithDefaultCredentials(rawString);
+    }
+
     protected String getSecret(String secretName)
     {
         return secretsManager.getSecret(secretName);
